[ruby/yarp] Move debug into its own file
https://github.com/ruby/yarp/commit/139362c90a
